Kafka.Internal.Response

Documentation

type Response a = Either Error a

A response from Kafka can either be a failure or the value that was expected.

data FetchResponse

Result of a single Kafka Fetch.

Constructors

FetchResponse

Fields

fetchMessages :: [ByteString]

List of messages returned in the response for the Fetch request.

fetchNewOffset :: !Offset

New offset at which the next Fetch request should start reading in the same topic and partition to access the messages that follow the messages returned in this response.

getFetchResponse :: Fetch -> Get FetchResponse

Parses a single FetchResponse for the given Fetch request.

getMultiFetchResponse :: [Fetch] -> Get [FetchResponse]

Parses FetchResponses for each of the given Fetch requests.

getOffsetsResponse :: Get [Offset]

Parses the response for an Offsets request.