Kafka.Internal.Transport

Documentation

class Transport t where

Types that provide a means to send and receive bytes.

Methods

send :: t -> ByteString -> IO ()

Send the given ByteString down the transport.

This must block until the request has been finished.

recv :: t -> Int -> IO ByteString

Read up to the given number of bytes from the stream.

The returned ByteString may be empty if the end of the stream was reached.

recvExactly :: Transport t => t -> Int -> IO ByteString

Keep reading from the given transport until the given number of bytes are read or the end of the stream is reached -- whichever comes first.

data Socket :: *

withConnection :: ByteString -> Int -> (Socket -> IO a) -> IO a

Open a connection, execute the given operation on it, and ensure it is closed afterwards even if an exception was thrown.

withConnection "localhost" 9092 $ \conn ->
   doStuff conn
   fail "something went wrong"

Throws an IOException if we were unable to open the connection.

connect :: ByteString -> Int -> IO Socket

Create a new connection.

Connects to the given hostname and port. Throws an IOException in case of failure.

close :: Socket -> IO ()

Close the socket. All future operations on the socket object will fail. The remote end will receive no more data (after queued data is flushed).